+++ packages/prose/src/routes/+layout.svelte
@@ -1,19 +1,19 @@
 <script lang="ts">
-	import ui from '@fat-fuzzy/ui';
-	import '$lib/styles/css/main.css';
+	import ui from '@fat-fuzzy/ui'
+	import '$lib/styles/css/main.css'
 
-	const { LayoutSidebar } = ui.content;
+	const {LayoutSidebar} = ui.content
 	// import { i18n } from '$lib/i18n';
 	// import { ParaglideJS } from '@inlang/paraglide-sveltekit';
-	let { children } = $props();
-	let path = '';
+	let {children} = $props()
+	let path = ''
 	let items = [
 		{
 			slug: 'entries',
 			title: 'Entries',
-			items: []
-		}
-	];
+			items: [],
+		},
+	]
 
 	let nav = {
 		path,
@@ -27,8 +27,8 @@
 		color: 'primary:600',
 		position: 'sticky',
 		place: 'left',
-		formaction: 'toggleSidebar'
-	};
+		formaction: 'toggleSidebar',
+	}
 </script>
 
 <!--