+++ packages/validation/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @fat-fuzzy/validation
 
+## 0.1.10
+
+### Patch Changes
+
+- d1028fd: Dynamic nested links, linting, config & dependency updates
+
 ## 0.1.9
 
 ### Patch Changes